Move new notebook command to core.
