+++ extensions/notebook/src/common/notebookUtils.ts
@@ -19,10 +19,6 @@ export class NotebookUtils {
 
 	constructor() { }
 
-	public async newNotebook(options?: azdata.nb.NotebookShowOptions): Promise<azdata.nb.NotebookEditor> {
-		return azdata.nb.showNotebookDocument(vscode.Uri.from({ scheme: 'untitled' }), options);
-	}
-
 	public async openNotebook(): Promise<void> {
 		try {
 			let filter: { [key: string]: Array<string> } = {};
