+++ src/sql/workbench/services/notebook/browser/notebookService.ts
@@ -17,13 +17,13 @@ import { NotebookChangeType, CellType } from 'sql/workbench/services/notebook/co
 import { IBootstrapParams } from 'sql/workbench/services/bootstrap/common/bootstrapParams';
 import { BaseTextEditor } from 'vs/workbench/browser/parts/editor/textEditor';
 import { Range } from 'vs/editor/common/core/range';
-import { IEditorPane } from 'vs/workbench/common/editor';
 import { INotebookInput } from 'sql/workbench/services/notebook/browser/interface';
 import { INotebookShowOptions } from 'sql/workbench/api/common/sqlExtHost.protocol';
 import { NotebookViewsExtension } from 'sql/workbench/services/notebook/browser/notebookViews/notebookViewsExtension';
 import { ICodeEditorViewState } from 'vs/editor/common/editorCommon';
 import { JUPYTER_PROVIDER_ID } from 'sql/workbench/common/constants';
 import { IStandardKernelWithProvider } from 'sql/workbench/services/notebook/browser/models/notebookUtils';
+import { IEditorPane } from 'vs/workbench/common/editor';
 
 export const SERVICE_ID = 'sqlNotebookService';
 export const INotebookService = createDecorator<INotebookService>(SERVICE_ID);
